Add files via upload
This is subVI I created to work with the C++ firmware found in this repository. It will get you started with your project, but it isn't a winner take all, as you may be 
trying to interface with different hardware, or you do not need to communicate with an ultrasonic sensor.